requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python
git+https://github.com/facebookresearch/detectron2.git
numpy